requirements.txt
customtkinter==5.2.2
faster-whisper==1.1.1
coqui-tts==0.27.5
coqpit-config==0.2.4
torch==2.5.1
torchaudio==2.5.1
numpy==1.26.4
scipy==1.15.3
librosa==0.11.0
numba==0.61.2
pydub==0.25.1
sounddevice==0.5.1
soundfile==0.13.1
noisereduce==3.0.3
requests==2.32.3
python-dotenv==1.0.1
webrtcvad==2.0.10
websocket-client==1.8.0
simpleaudio==1.0.4
transformers>=4.57,<5
tokenizers>=0.22,<0.23
huggingface-hub>=0.34,<1
protobuf>=5.26.1,<7
pillow<12
Unidecode==1.3.8
eng-to-ipa==0.0.2
pypinyin==0.54.0
jieba==0.42.1
cn2an==0.5.23
wavmark==0.0.3
